requirements.txt
flask==3.0.2
psycopg2-binary
requests==2.32.3
bs4
chardet
pytz
Gunicorn==20.1.0
selenium==4.23.1
mysqlclient==2.2.4
sqlalchemy==2.0.32
flask_sqlalchemy
python-dotenv
APScheduler==3.10.4
Flask-APScheduler==1.13.1
pymongo==4.8.0
selenium-wire==5.1.0
webdriver-manager
pika
flask-cors
openai
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-api-python-client==2.169.0
google-genai==1.13.0
tenacity==9.0.0
Werkzeug
langchain-openai
langchain-google-genai==2.0.10
browser_use==0.1.40
flask[async]
